// src/types/bindings/ReportMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ReportMetadata = { report_name: string, report_version: number, generated_at: string, execution_time_ms: bigint, filter_hash: string, total_records: number, total_pages: number, page: number, page_size: number, };